[WFCORE-2037] Elytron SupportLevel has now moved to package 'org.wildfly.security.auth'.
